refactor: Change folder structure and make loader component generic
