+++ src/components/Layout/index.tsx
@@ -5,7 +5,7 @@ import Footer from '@/components/footer';
 import styles from '@/components/Layout/Layout.module.scss';
 import NavBar from '@/components/navBar';
 import { useGetUserQuery } from '@/app/services/userApi';
-import { Loader } from '../tasks/card/Loader';
+import { Loader } from '../Loaders/Loader';
 
 interface Props {
     children?: ReactNode;
